store, image, cmd: make 'snap download' leave partials
Without this change, a network error during a 'snap download' will
remove the .partial file.

While this is probably OK for snapd, as it's the simpler & safer
route¹, it's not OK for 'snap download'; the .partial should be left
so that a following 'snap download' will resume, as the user can see
and remove the file if needed.

So, this change adds a LeavePartialOnError to store.DownloadOptions,
exports it through image.DownloadOptions, and sets that to true in
cmd_download.

Also note, it explicitly makes sure it doesn't leave _empty_ partial
files, as those would just be noise.

----

1. We might want to revisit what we do in snapstate so that there also
we can resume, but it needs careful consideraiton of available space
and removing stale partial files, which is one of the hard problems.

+++ store/store.go
@@ -1334,8 +1334,9 @@ func (e HashError) Error() string {
 }
 
 type DownloadOptions struct {
-	RateLimit     int64
-	IsAutoRefresh bool
+	RateLimit           int64
+	IsAutoRefresh       bool
+	LeavePartialOnError bool
 }
 
 // Download downloads the snap addressed by download info and returns its
@@ -1375,11 +1376,21 @@ func (s *Store) Download(ctx context.Context, name string, targetPath string, do
 		return err
 	}
 	defer func() {
+		fi, _ := w.Stat()
 		if cerr := w.Close(); cerr != nil && err == nil {
 			err = cerr
 		}
 		if err != nil {
-			os.Remove(w.Name())
+			doRemove := dlOpts == nil || !dlOpts.LeavePartialOnError
+			if !doRemove {
+				// also do remove if the file is empty
+				if fi != nil && fi.Size() == 0 {
+					doRemove = true
+				}
+			}
+			if doRemove {
+				os.Remove(w.Name())
+			}
 		}
 	}()
 	if resume > 0 {
